➖: Uninstall @typescript-eslint/{eslint-plugin,parser} (#74)
## 🤔 What was the reason for the change

Uninstall following packages from devDependencies because they will be
automatically installed by npm (peerDependencies of
`eslint-config-universe`).
* `@typescript-eslint/eslint-plugin`
* `@typescript-eslint/parser`
* `prettier`

With the major version update of Prettier, it became necessary to also
perform a major version update for `eslint-plugin-prettier`. However,
there was an issue with the version specification in
`eslint-config-universe`, which caused a problem where only Prettier
would undergo a major version update if left unchanged. To address this,
we specified in the overrides that we would not perform a major version
update for Prettier.